backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
redis==5.1.1
django-redis==5.4.0
gunicorn==23.0.0
Pillow==10.4.0
python-decouple==3.8
dj-database-url==2.2.0
